requirements.txt
fastapi
uvicorn
httpx
pyyaml
requests

